MNT: implement some suggestions
