+++ pcdscalc/tests/test_diffraction.py
@@ -6,19 +6,6 @@ from pcdscalc import diffraction
 logger = logging.getLogger(__name__)
 
 
-@pytest.mark.parametrize('energy, expected', [
-                         pytest.param(2e3, 6.1992e-10),
-                         pytest.param(10e3, 1.23984e-10),
-                         pytest.param(9e3, 1.3776e-10),
-                         pytest.param(8e3, 1.5498e-10)
-                         ])
-def test_energy_to_wavelength(energy, expected):
-    # tested agains the old code
-    res = diffraction.energy_to_wavelength(energy)
-    logger.debug('Expected: %s Received: %s', expected, res)
-    assert np.isclose(expected, res)
-
-
 @pytest.mark.parametrize('energy, material, reflection, expected', [
                          pytest.param(10e3, 'Si', (1, 1, 1),
                                       (11.402710639982848, 713.4828146545175)),
@@ -47,19 +34,6 @@ def test_get_geometry(energy, material, reflection, expected):
 def test_bragg_angle(material_id, hkl, energy, expected):
     # tested agains the old code
     res = diffraction.bragg_angle(material_id, hkl, energy)
-    logger.debug('Expected: %s Received: %s', expected, res)
-    assert np.isclose(expected, res)
-
-
-@pytest.mark.parametrize('wavelength, expected', [
-                         pytest.param(100, 1.23984e-08),
-                         pytest.param(9, 1.3775999999999998e-07),
-                         pytest.param(7, 1.7712e-07),
-                         pytest.param(0.8, 1.5498e-06)
-                         ])
-def test_wavelength_to_energy(wavelength, expected):
-    # tested agains the old code
-    res = diffraction.wavelength_to_energy(wavelength)
     logger.debug('Expected: %s Received: %s', expected, res)
     assert np.isclose(expected, res)
